+++ src/state_flow/state.clj
@@ -2,7 +2,6 @@
   (:refer-clojure :exclude [eval get])
   (:require [cats.context :as ctx :refer [*context*]]
             [cats.core :as m]
-            [cats.data :as d]
             [cats.monad.exception :as e]
             [cats.monad.state :as state]
             [cats.protocols :as p]
@@ -10,47 +9,56 @@
 
 (declare error-context)
 
+(defn- result-or-err [f & args]
+  (let [result ((e/wrap (partial apply f)) args)]
+    (if (e/failure? result)
+      result
+      @result)))
+
+(defn error-state [mfn]
+  (state/state
+   (fn [s]
+     (let [new-pair ((e/wrap mfn) s)]
+       (if (e/failure? new-pair)
+         [new-pair s]
+         @new-pair)))
+   error-context))
+
 (def error-context
   "Same as state monad context, but short circuits if error happens, place error in return value"
   (reify
     p/Context
 
-    p/Extract
-    (-extract [mv] (p/-extract mv))
-
     p/Functor
-    (-fmap [_ f fv]
-      (state/state (fn [s]
-                     (let [[v ns]  ((p/-extract fv) s)]
-                       (d/pair (f v) ns)))
-                   error-context))
+    (-fmap [self f fv]
+      (error-state
+       (fn [s]
+         (let [[v s'] ((p/-extract fv) s)]
+           (if (e/failure? v)
+             [v s']
+             [(result-or-err f v) s'])))))
 
     p/Monad
     (-mreturn [_ v]
-      (state/state (partial d/pair v) error-context))
+      (error-state #(vector v %)))
 
     (-mbind [_ self f]
-      (state/state (fn [s]
-                     (let [mp ((e/wrap (p/-extract self)) s)]
-                       (if (e/failure? mp)
-                         (d/pair mp s)
-                         (if (e/failure? (.-fst @mp))
-                           @mp
-                           (let [new-pair ((e/wrap (p/-extract (f (.-fst @mp)))) (.-snd @mp))]
-                             (if (e/success? new-pair)
-                               @new-pair
-                               (d/pair new-pair (.-snd @mp))))))))
-                   error-context))
+      (error-state
+       (fn [s]
+         (let [[v s'] ((p/-extract self) s)]
+           (if (e/failure? v)
+             [v s']
+             ((p/-extract (f v)) s'))))))
 
     state/MonadState
     (-get-state [_]
-      (state/state #(d/pair %1 %1) error-context))
+      (error-state #(vector %1 %1)))
 
     (-put-state [_ newstate]
-      (state/state #(d/pair % newstate) error-context))
+      (error-state #(vector % newstate)))
 
     (-swap-state [_ f]
-      (state/state #(d/pair %1 (f %1)) error-context))
+      (error-state #(vector %1 (f %1))))
 
     p/Printable
     (-repr [_]
@@ -64,9 +72,9 @@
   (state/get error-context))
 
 (defn gets
-  [f]
-  "Returns the equivalent of (fn [state] [state, (f state)])"
-  (state/gets f))
+  [f & args]
+  "Returns the equivalent of (fn [state] [state, (apply f state args)])"
+  (state/gets #(apply f % args) error-context))
 
 (defn put
   "Returns the equivalent of (fn [state] [state, new-state])"
@@ -74,9 +82,9 @@
   (state/put new-state error-context))
 
 (defn modify
-  "Returns the equivalent of (fn [state] [state, (swap! state f)])"
-  [f]
-  (state/swap f error-context))
+  "Returns the equivalent of (fn [state] [state, (apply swap! state f args)])"
+  [f & args]
+  (state/swap #(apply f % args) error-context))
 
 (defn return
   "Returns the equivalent of (fn [state] [v, state])"
@@ -91,11 +99,18 @@
 (defn wrap-fn
   "Wraps a (possibly side-effecting) function to a state monad"
   [my-fn]
-  (state/state (fn [s]
-                 (d/pair (my-fn) s))
-               error-context))
+  (error-state (fn [s] [(my-fn) s])))
 
 (def state? state/state?)
 (def run state/run)
 (def eval state/eval)
 (def exec state/exec)
+
+(defn ensure-step
+  "Internal use only.
+
+  Given a state-flow step, returns value as/is, else wraps value in a state-flow step."
+  [value]
+  (if (state? value)
+    value
+    (return value)))