+++ test/state_flow/cljtest_test.clj
@@ -2,8 +2,9 @@
   (:require [clojure.test :as t :refer [deftest testing is]]
             [matcher-combinators.test :refer [match?]]
             [matcher-combinators.matchers :as matchers]
+            [state-flow.assertions.matcher-combinators :as assertions.matcher-combinators]
             [state-flow.test-helpers :as test-helpers :refer [this-line-number]]
-            [state-flow.cljtest :as cljtest :refer [defflow]]
+            [state-flow.cljtest :refer [defflow]]
             [state-flow.core :as state-flow :refer [flow]]
             [state-flow.state :as state]))
 
@@ -12,23 +13,22 @@
 (deftest test-match?
   (testing "passing cases"
     (testing "with literals for expected and actual"
-      (let [[ret state] (state-flow/run (cljtest/match? "DESC" 3 3) {:initial :state})]
+      (let [[ret state] (state-flow/run (testing "DESC" (assertions.matcher-combinators/match? 3 3)) {:initial :state})]
         (testing "returns actual (literal)"
           (is (= 3 ret)))
         (testing "doesn't change state"
           (is (= {:initial :state} state)))))
 
     (testing "with state monad for actual"
-      (let [[ret state] (state-flow/run (cljtest/match? "DESC" test-helpers/add-two 3) {:value 1})]
+      (let [[ret state] (state-flow/run (testing "DESC" (assertions.matcher-combinators/match? 3 test-helpers/add-two)) {:value 1})]
         (testing "returns actual (derived from state)"
           (is (= 3 ret)))
         (testing "doesn't change state"
           (is (= {:value 1} state)))))
 
     (testing "with explicit matcher for expected"
-      (let [[ret state] (state-flow/run (cljtest/match? "DESC"
-                                                        test-helpers/add-two
-                                                        (matchers/equals 3)) {:value 1})]
+      (let [[ret state] (state-flow/run (testing "DESC" (assertions.matcher-combinators/match? (matchers/equals 3)
+                                                                                               test-helpers/add-two)) {:value 1})]
         (testing "returns actual (derived from state)"
           (is (= 3 ret)))
         (testing "doesn't change state"
@@ -39,8 +39,8 @@
             (test-helpers/run-flow
              (flow "flow"
                (test-helpers/delayed-add-two 100)
-               (cljtest/match? "2" get-value-state 2 {:times-to-try 2
-                                                      :sleep-time 75}))
+               (testing "2" (assertions.matcher-combinators/match? 2 get-value-state {:times-to-try 2
+                                                                                      :sleep-time 110})))
              {:value (atom 0)})]
         (testing "returns actual (derived from state)"
           (is (= 2 flow-ret))))))
@@ -50,10 +50,9 @@
       (let [three-lines-before-call-to-match (this-line-number)
             {:keys [flow-ret flow-state report-data]}
             (test-helpers/run-flow
-             (cljtest/match? "contains with monadic left value"
-                             (state/gets :value)
-                             (matchers/equals {:n 1})
-                             {:times-to-try 2})
+             (testing "contains with monadic left value" (assertions.matcher-combinators/match? (matchers/equals {:n 1})
+                                                                                                (state/gets :value)
+                                                                                                {:times-to-try 2}))
              {:value {:n 2}})]
         (testing "returns actual"
           (is (= {:n 2} flow-ret)))
@@ -69,8 +68,8 @@
             (test-helpers/run-flow
              (flow "flow"
                (test-helpers/delayed-add-two 200)
-               (cljtest/match? "2" get-value-state 2 {:times-to-try 2
-                                                      :sleep-time 75}))
+               (testing "2" (assertions.matcher-combinators/match? 2 get-value-state {:times-to-try 2
+                                                                                      :sleep-time 75})))
              {:value (atom 0)})]
         (testing "returns actual (derived from state)"
           (is (= 0 flow-ret)))
@@ -83,50 +82,51 @@
 ;; in the deftest below causes test failures. I think it has to do with calling macroexpand
 ;; within a macro body.
 (def flow-with-defaults
-  (macroexpand-1 '(defflow my-flow (cljtest/match? "equals" 1 1))))
+  (macroexpand-1 '(defflow my-flow (testing "equals" (assertions.matcher-combinators/match? 1 1)))))
 (def flow-with-optional-args
-  (macroexpand-1 '(defflow my-flow {:init (constantly {:value 1})} (cljtest/match? "equals" 1 1))))
+  (macroexpand-1 '(defflow my-flow {:init (constantly {:value 1})} (testing "equals" (assertions.matcher-combinators/match? 1 1)))))
 (def flow-with-binding-and-match
   (macroexpand-1 '(defflow my-flow {:init (constantly {:value 1
                                                        :map {:a 1 :b 2}})}
                     [value (state/gets :value)]
-                    (cljtest/match? value 1)
-                    (cljtest/match? (state/gets :map) {:b 2}))))
+                    (testing "1" (assertions.matcher-combinators/match? 1 value))
+                    (testing "b is 2" (assertions.matcher-combinators/match? {:b 2} (state/gets :map))))))
 
 (deftest test-defflow
   (testing "defines flow with default parameters"
     (is (= '(clojure.test/deftest
               my-flow
               (state-flow.core/run*
-                {}
-                (state-flow.core/flow "my-flow" (cljtest/match? "equals" 1 1))))
+               {}
+               (state-flow.core/flow "my-flow" (testing "equals" (assertions.matcher-combinators/match? 1 1)))))
            flow-with-defaults)))
 
   (testing "defines flow with optional parameters"
     (is (= '(clojure.test/deftest
               my-flow
               (state-flow.core/run*
-                {:init (constantly {:value 1})}
-                (state-flow.core/flow "my-flow" (cljtest/match? "equals" 1 1))))
+               {:init (constantly {:value 1})}
+               (state-flow.core/flow "my-flow" (testing "equals" (assertions.matcher-combinators/match? 1 1)))))
            flow-with-optional-args)))
 
   (testing "defines flow with binding and flow inside match?"
     (is (= '(clojure.test/deftest
               my-flow
               (state-flow.core/run*
-                {:init (constantly {:map {:a 1 :b 2} :value 1})}
-                (state-flow.core/flow
-                  "my-flow"
-                  [value (state/gets :value)]
-                  (cljtest/match? value 1)
-                  (cljtest/match? (state/gets :map) {:b 2}))))
+               {:init (constantly {:map {:a 1 :b 2} :value 1})}
+               (state-flow.core/flow
+                "my-flow"
+                 [value (state/gets :value)]
+                 (testing "1" (assertions.matcher-combinators/match? 1 value))
+                 (testing "b is 2" (assertions.matcher-combinators/match? {:b 2} (state/gets :map))))))
+
            flow-with-binding-and-match))))
 
 (defflow my-flow {:init (constantly {:value 1
                                      :map   {:a 1 :b 2}})}
   [value (state/gets :value)]
-  (cljtest/match? "1" value 1)
-  (cljtest/match? "b is 2" (state/gets :map) {:b 2}))
+  (testing "1" (assertions.matcher-combinators/match? 1 value))
+  (testing "b is 2" (assertions.matcher-combinators/match? {:b 2} (state/gets :map))))
 
 (deftest run-a-flow
   (is (match? {:value 1