+++ CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+## [2.2.6]
+
+- Fix issues with exceptions being thrown and not returned as the left value of
+  the error-state monad
+
+## [2.2.5]
+
+- add shell script to refactor match? expressions and `:require [state-flow.cljtest} ...`
+
+## [2.2.4]
+
+- state-flow.state/modify and state-flow.state/gets pass additional args to f
+- Introduce `state-flow.assertions.matcher-combinators/match?`
+  - deprecate `state-flow.cljtest/match?`
+  - add `state-flow.refactoring-tools.refactor-match/refactor-all` to help with
+    refactoring to the new version
+
+## [2.2.3]
+
+- Revert changes in `2.2.2` until a few issues are resolved
+
+## ~[2.2.2]~
+
+DO NOT USE VERSION 2.2.2
+
+Changes were reintroduced in `2.2.4`, so use that.
+
+## [2.2.1]
+- Use vectors in internal state data structure instead of cats pairs
+
+## [2.2.0]
+
+- Remove delay from the first try of `probe`
+- Refactor probe and change return value to include the probed value and check result of each try
+
 ## [2.1.5]
 
 - Make `state-flow.state/return` constructable/def'able outside monadic context.